app/requirements.txt
Flask-Migrate==2.5.3
Flask-Login==0.5.0
Flask-Babel==1.0.0
